+++ tests/files/c_source/multiplex_2.c
@@ -112,7 +112,7 @@ static inline uint32_t unpack_right_shift_u32(
     return (uint32_t)((uint32_t)(value & mask) >> shift);
 }
 
-ssize_t multiplex_2_shared_pack(
+int multiplex_2_shared_pack(
     uint8_t *dst_p,
     const struct multiplex_2_shared_t *src_p,
     size_t size)
@@ -289,7 +289,7 @@ bool multiplex_2_shared_s2_is_in_range(int8_t value)
     return (true);
 }
 
-ssize_t multiplex_2_normal_pack(
+int multiplex_2_normal_pack(
     uint8_t *dst_p,
     const struct multiplex_2_normal_t *src_p,
     size_t size)
@@ -423,7 +423,7 @@ bool multiplex_2_normal_s2_is_in_range(int8_t value)
     return (true);
 }
 
-ssize_t multiplex_2_extended_pack(
+int multiplex_2_extended_pack(
     uint8_t *dst_p,
     const struct multiplex_2_extended_t *src_p,
     size_t size)
@@ -774,7 +774,7 @@ bool multiplex_2_extended_s7_is_in_range(int32_t value)
     return ((value >= -8388608) && (value <= 8388607));
 }
 
-ssize_t multiplex_2_extended_types_pack(
+int multiplex_2_extended_types_pack(
     uint8_t *dst_p,
     const struct multiplex_2_extended_types_t *src_p,
     size_t size)
